requirements.txt
Pillow>=10.0
PyYAML>=6.0
requests>=2.31
